Enviado

+++ CrudPHP/excluir.php
@@ -1,11 +1,23 @@
 <?php
-require "banco.php";
+session_start(); 
+if (!isset($_SESSION['logado']) || $_SESSION['logado'] !== true) {
+    header('Location: index.php');
+    exit;
+}
 
-$id = $_GET["id"];
+require "banco.php"; 
+$id = $_GET["id"] ?? null;
+
+
+if (!$id) {
+    header("Location: lista.php");
+    exit;
+}
 
 $sql = $pdo->prepare("DELETE FROM usuarios WHERE id = :id");
 $sql->bindParam(":id", $id);
 $sql->execute();
 
-header("Location: lista.php"); // redireciona
+header("Location: lista.php"); /
 exit;
+?>
